# [FIXED] Changes to Varz content and fixed race conditions