@kozlovic 2019-05-09
Here are some benchmark results between master (old) and this branch (new), for Varz - when invoked programmatically, which requires deep copy, and VarzHttp which is when getting it from http endpoint.
```
benchmark                 old ns/op     new ns/op     delta
Benchmark_Varz-16         3339665       3338554       -0.03%
Benchmark_VarzHttp-16     6578618       5035424       -23.46%

benchmark                 old allocs     new allocs     delta
Benchmark_Varz-16         113            113            +0.00%
Benchmark_VarzHttp-16     269            266            -1.12%

benchmark                 old bytes     new bytes     delta
Benchmark_Varz-16         42412         42864         +1.07%
Benchmark_VarzHttp-16     57436         58345         +1.58%
```